@@ -1,30 +1,45 @@
-import React, { Suspense } from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { Suspense } from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Skeleton } from "antd";
 
-import { HeaderBar } from "./components/header";
-
 import { MAIN_ROUTES } from "./route";
+import { TopFrame } from "./components/topframe";
 
 import "antd/dist/antd.css";
 
 function App() {
   return (
-    <BrowserRouter basename="/view">
-      {/* <HeaderBar /> */}
-      <Suspense fallback={<Skeleton active={true} />}>
-        <Switch>
-          {MAIN_ROUTES.map((route) => {
-            const Component = route.component;
-            return (
-              <Route path={route.path} exact={route.exact} key={route.key}>
-                <Component />
-              </Route>
-            );
-          })}
-        </Switch>
-      </Suspense>
-      {/* <Redirect to="/home" /> */}
+    <BrowserRouter>
+      <TopFrame navigation={MAIN_ROUTES} />
+      <main className="flex w-full">
+        <Suspense fallback={<Skeleton active={true} />}>
+          <Routes>
+            {MAIN_ROUTES.map((route) => {
+              const Component = route.component();
+              return (
+                <Route
+                  path={route.path}
+                  key={route.key}
+                  element={<Component />}
+                >
+                  {route.nestedRoute
+                    ? route.nestedRoute.map((route) => {
+                        const Element = route.component();
+                        return (
+                          <Route
+                            path={route.path}
+                            key={route.path}
+                            element={<Element />}
+                          />
+                        );
+                      })
+                    : null}
+                </Route>
+              );
+            })}
+          </Routes>
+        </Suspense>
+      </main>
     </BrowserRouter>
   );
 }
