@@ -11,7 +11,7 @@ impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
         vec![
             Box::new(m20220101_000001_create_table::Migration),
-            Box::new(m20220101_000002_create_table::Migration),
+            // Box::new(m20220101_000002_create_table::Migration),
             // Box::new(m20220101_000003_setup::Migration)
         ]
     }
