@@ -1,2 +1,2 @@
-declare module '*.css';
-declare module '*.module.css';
+declare module "*.css";
+declare module "*.module.css";
