@@ -1,11 +1,12 @@
 import React from "react";
-
-import styles from "./home.module.css";
+import { CommandNode } from "../components/node/step/command";
+import { Studio } from "../components/studio";
 
 export function Homepage() {
   return (
-    <div className={styles.container}>
-      <h1>Test Suite page coming soon</h1>
-    </div>
+    // <div className="flex justify-center">
+    //   <h1>Test Suite page coming soon</h1>
+    // </div>
+    <Studio />
   );
 }
